Move perf counter and thread bits to PerfCounterHelper
Next up: adding some worker pool stats to a few exceptions.
